[NET][USR] Added intermodule communication (#82)
* [NET][USR] Added intermodule communication

* [ACE] Restructured disconnect mechanics

* [NET][WBG] Remade game status mechanics

+++ gameNetworking/implementations/game_consumer_impl/disconnect.py
@@ -1,11 +1,11 @@
 from django.conf import settings
+from channels.db import database_sync_to_async
 
 from customUser.models.queries import create_game_archive
 
-from ...models.queries import delete_game, delete_game_user
-from ...scheduler.scheduler import add_delayed_task
-from ...scheduler.scheduler import remove_delayed_task
-from ...scheduler.scheduler import delete_states_queue
+from ...enums import GameState
+from ...models.queries import delete_game
+from ...scheduler.scheduler import *
 
 
 class Disconnector:
@@ -14,14 +14,14 @@ class Disconnector:
         self._consumer = consumer
 
     async def disconnect(self):
+        await self._consumer.refresh_game()
         game = self._consumer.get_game()
+        player = self._consumer.get_game_user()
+        await self._clear_in_game_status(player)
         if self._consumer.closed_after_disconnect():
-            await self._consumer.refresh_game()
-            player = self._consumer.get_game_user()
-            game = self._consumer.get_game()
             if game:
                 await player.backup(self._consumer)
-                if not game.is_backuped:
+                if check_game_state(str(game.id)) != GameState.BACKUPED:
                     await self._consumer.send_message_to_opponent(
                         {}, "opponent_disconnect"
                     )
@@ -32,19 +32,20 @@ class Disconnector:
                         settings.DELETE_GAME_TIMEOUT,
                         settings.DELETE_GAME_TIMEOUT_FUNC
                     )
-            else:
-                await delete_game_user(player.id)
-            return
-        winner = self._consumer.get_winner()
-        await self._consumer.send_message_to_opponent(
-            {"winner" : winner},
-            "game_end"
-        )
-        await create_game_archive(game, winner)
-        await self._remove_player_from_group(game.id)
-        await delete_game(game.id)
-        self._remove_all_gameplay_tasks()
-        delete_states_queue(game.id)
+        else:
+            winner = self._consumer.get_winner()
+            await self._remove_player_from_group(game.id)
+            if check_game_state(str(game.id)) != GameState.DELETED:
+                update_game_state(str(game.id), GameState.DELETED)
+                await database_sync_to_async(create_game_archive)(game, winner)
+                await delete_game(game.id)
+                self._remove_all_gameplay_tasks()
+                delete_player_states_queue(game.id)
+                add_delayed_task(
+                    f'limit_game_state_lifetime_{game.id}',
+                    settings.DELETE_GAME_STATE_TIMEOUT,
+                    settings.DELETE_GAME_STATE_TIMEOUT_FUNC
+                )
 
     async def _remove_player_from_group(self, game_id):
         await self._consumer.channel_layer.group_discard(
@@ -60,3 +61,7 @@ class Disconnector:
         remove_delayed_task(f'limit_hub_time_{g_u_id}')
         remove_delayed_task(f'limit_hub_time_{opp_id}')
         remove_delayed_task(f'limit_opponent_rejoin_time_{g_u_id}')
+
+    async def _clear_in_game_status(self, player):
+        player_user = await player.get_user()
+        await database_sync_to_async(player_user.clear_in_game)()
